Scenes: Refactor original snapshot button in a new component
