+++ public/app/features/dashboard-scene/scene/DashboardScene.tsx
@@ -1,10 +1,8 @@
-import { css } from '@emotion/css';
 import * as H from 'history';
-import React from 'react';
 import { Unsubscribable } from 'rxjs';
 
-import { AppEvents, CoreApp, DataQueryRequest, NavIndex, NavModelItem, locationUtil, textUtil } from '@grafana/data';
-import { locationService, config } from '@grafana/runtime';
+import { AppEvents, CoreApp, DataQueryRequest, NavIndex, NavModelItem, locationUtil } from '@grafana/data';
+import { locationService } from '@grafana/runtime';
 import {
   getUrlSyncManager,
   SceneFlexLayout,
@@ -23,7 +21,6 @@ import {
   VizPanel,
 } from '@grafana/scenes';
 import { Dashboard, DashboardLink } from '@grafana/schema';
-import { ConfirmModal } from '@grafana/ui';
 import appEvents from 'app/core/app_events';
 import { LS_PANEL_COPY_KEY } from 'app/core/constants';
 import { getNavModel } from 'app/core/selectors/navModel';
@@ -33,7 +30,7 @@ import { DashboardModel } from 'app/features/dashboard/state';
 import { dashboardWatcher } from 'app/features/live/dashboard/dashboardWatcher';
 import { VariablesChanged } from 'app/features/variables/types';
 import { DashboardDTO, DashboardMeta, SaveDashboardResponseDTO } from 'app/types';
-import { ShowModalReactEvent, ShowConfirmModalEvent } from 'app/types/events';
+import { ShowConfirmModalEvent } from 'app/types/events';
 
 import { PanelEditor } from '../panel-edit/PanelEditor';
 import { SaveDashboardDrawer } from '../saving/SaveDashboardDrawer';
@@ -492,47 +489,6 @@ export class DashboardScene extends SceneObjectBase<DashboardSceneState> {
       console.error('Failed to star dashboard', err);
     }
   }
-
-  public onOpenSnapshotOriginalDashboard = () => {
-    // @ts-ignore
-    const relativeURL = this.getInitialSaveModel()?.snapshot?.originalUrl ?? '';
-    const sanitizedRelativeURL = textUtil.sanitizeUrl(relativeURL);
-    try {
-      const sanitizedAppUrl = new URL(sanitizedRelativeURL, config.appUrl);
-      const appUrl = new URL(config.appUrl);
-      if (sanitizedAppUrl.host !== appUrl.host) {
-        appEvents.publish(
-          new ShowModalReactEvent({
-            component: ConfirmModal,
-            props: {
-              title: 'Proceed to external site?',
-              modalClass: css({
-                width: 'max-content',
-                maxWidth: '80vw',
-              }),
-              body: (
-                <>
-                  <p>
-                    {`This link connects to an external website at`} <code>{relativeURL}</code>
-                  </p>
-                  <p>{"Are you sure you'd like to proceed?"}</p>
-                </>
-              ),
-              confirmVariant: 'primary',
-              confirmText: 'Proceed',
-              onConfirm: () => {
-                window.location.href = sanitizedAppUrl.href;
-              },
-            },
-          })
-        );
-      } else {
-        locationService.push(sanitizedRelativeURL);
-      }
-    } catch (err) {
-      console.error('Failed to open original dashboard', err);
-    }
-  };
 
   public onOpenSettings = () => {
     locationService.partial({ editview: 'settings' });
