+++ public/app/features/dashboard-scene/scene/NavToolbarActions.tsx
@@ -15,6 +15,7 @@ import { DashboardInteractions } from '../utils/interactions';
 import { dynamicDashNavActions } from '../utils/registerDynamicDashNavAction';
 
 import { DashboardScene } from './DashboardScene';
+import { GoToSnapshotOriginButton } from './GoToSnapshotOriginButton';
 
 interface Props {
   dashboard: DashboardScene;
@@ -87,13 +88,9 @@ export function ToolbarActions({ dashboard }: Props) {
     group: 'icon-actions',
     condition: meta.isSnapshot && !isEditing,
     render: () => (
-      <ToolbarButton
-        key="button-snapshot"
-        tooltip={t('dashboard.toolbar.open-original', 'Open original dashboard')}
-        icon="link"
-        onClick={() => {
-          dashboard.onOpenSnapshotOriginalDashboard();
-        }}
+      <GoToSnapshotOriginButton
+        // @ts-ignore
+        originalURL={dashboard.getInitialSaveModel()?.snapshot?.originalUrl ?? ''}
       />
     ),
   });
